agenticlearning/12_genie_langchain_whl/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "genie_langchain_agent"
version = "0.1.0"
description = "Simple Genie Agent with LangChain — packaged as a wheel"
requires-python = ">=3.9"
dependencies = [
    "langchain",
    "langchain-core",
    "langchain-databricks",
    "mlflow>=3",
    "databricks-agents",
    "pyspark",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
